+++ src/ytsk/config.py
@@ -1,0 +1,8 @@
+from pathlib import Path
+
+
+ROOT_DIR = Path(__file__).parents[2]
+
+DATA_DIR = ROOT_DIR / "data"
+EXTERNAL_DIR = DATA_DIR / "external"
+RAW_DIR = DATA_DIR / "raw"